pyproject.toml
[project]
name = "futures-langchain"
version = "0.1.0"
description = "期货多智能体分析系统 - 基于LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.15.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=3.0.3",
    "psycopg-binary>=3.3.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.0.0",
    "zhipuai>=2.1.0",
    "matplotlib>=3.9.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "ta>=0.11.0",
    "langchain-deepseek>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
